# openapi.yaml
openapi: 3.0.0
info:
  title: Dokany API
  description: منصة دكاني - واجهات برمجة التطبيقات
  version: 1.1.0
  contact:
    name: Dokany Team
    url: https://dokany.com
servers:
  - url: https://api.dokany.com
    description: Production Server
  - url: http://localhost:3000
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    storeIdHeader:
      type: apiKey
      in: header
      name: x-store-id
  schemas:
...
    CouponValidationRequest:
      type: object
      properties:
        storeId:
          type: string
        code:
          type: string
        cartTotalAmount:
          type: string
        customerId:
          type: string
        categoryId:
          type: string
        productId:
          type: string
    HaggleSession:
      type: object
      properties:
        id:
          type: string
        sessionCode:
          type: string
        status:
          type: string
          enum: [active, counter_offered, accepted, rejected, expired, cancelled]
        currentOffer:
          type: string
...
security:
  - bearerAuth: []
  - storeIdHeader: []
paths:
...
  /coupons/validate:
    post:
      tags: [Coupons]
      summary: التحقق من كوبون الخصم
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponValidationRequest'
      responses:
        '200':
          description: الكوبون صالح
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  couponId: { type: string }
                  calculatedDiscountAmount: { type: string }
                  finalAmount: { type: string }
        '400':
          description: الكوبون غير صالح
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean, example: false }
                  error: { type: string }

  # ============================================================
  # 💬 HAGGLE
  # ============================================================
  /haggle:
    post:
      tags: [Haggle]
      summary: إنشاء جلسة مساومة
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storeId: { type: string }
                productId: { type: string }
                originalPrice: { type: string }
                minAllowedPrice: { type: string }
                initialOffer: { type: string }
                maxRounds: { type: integer }
                durationInMinutes: { type: integer }
                strategy: { type: string, enum: [aggressive, friendly, middle_ground] }
      responses:
        '201':
          description: تم إنشاء الجلسة
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  session: { $ref: '#/components/schemas/HaggleSession' }
                  botMessage: { type: string }

    patch:
      tags: [Haggle]
      summary: تقديم عرض في جلسة مساومة
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId: { type: string }
                offeredPrice: { type: string }
                message: { type: string }
      responses:
        '200':
          description: تم تقديم العرض
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  session: { $ref: '#/components/schemas/HaggleSession' }
                  botMessage: { type: string }
...